fix: 사용자의 유형에 따라 AppBar의 title을 다르게 설정하도록 수정 및 사용자 유형을 매개변수로 추가
